+++ Helper/FileEncoding.php
@@ -12,6 +12,7 @@ namespace Inriver\Adapter\Helper;
 
 use Magento\Framework\Filesystem\Driver\File as FileDriver;
 
+//phpcs:ignore Magento2.Exceptions.TryProcessSystemResources.MissingTryCatch
 use function stream_copy_to_stream;
 use function unlink;