+++ Model/Import/Product/Type/Configurable.php
@@ -158,7 +158,7 @@ class Configurable extends \Magento\ConfigurableImportExport\Model\Import\Produc
                     $quotedChildren = $this->connection->quoteInto('IN (?)', $deleteRelation);
                     $this->connection->delete(
                         $mainTable,
-                        'product_id {' . $quoted . '} AND attribute_id {' . $quotedChildren . '}'
+                        'product_id ' . $quoted . ' AND attribute_id ' . $quotedChildren
                     );
                 }
             }
@@ -193,7 +193,7 @@ class Configurable extends \Magento\ConfigurableImportExport\Model\Import\Produc
                     $quotedChildren = $this->connection->quoteInto('IN (?)', $deleteRelation);
                     $this->connection->delete(
                         $linkTable,
-                        'parent_id {' . $quoted . '} AND product_id {' . $quotedChildren . '}'
+                        'parent_id ' . $quoted . ' AND product_id ' . $quotedChildren
                     );
                 }
             }
@@ -228,7 +228,7 @@ class Configurable extends \Magento\ConfigurableImportExport\Model\Import\Produc
                     $quotedChildren = $this->connection->quoteInto('IN (?)', $deleteRelation);
                     $this->connection->delete(
                         $relationTable,
-                        'parent_id {' . $quoted . '} AND child_id {' . $quotedChildren . '}'
+                        'parent_id ' . $quoted . ' AND child_id ' . $quotedChildren
                     );
                 }
             }